Move searchGooglePlay to apiHelper for easier testing Issue: #6
